# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"物资库存管理系统 API","description":"SpringBoot + MySQL 版 RESTful 接口文档","version":"1.0.0"},"servers":[{"url":"http://localhost:3000","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/users/{id}":{"put":{"tags":["user-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserView"}}}}}},"delete":{"tags":["user-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/materials/{id}":{"get":{"tags":["material-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Material"}}}}}},"put":{"tags":["material-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Material"}}}}}},"delete":{"tags":["material-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"list","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultUserView"}}}}}},"post":{"tags":["user-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserView"}}}}}}},"/api/outbound":{"get":{"tags":["outbound-controller"],"operationId":"list_1","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"materialId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultMapStringObject"}}}}}},"post":{"tags":["outbound-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OutboundRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/outbound/{id}/reject":{"post":{"tags":["outbound-controller"],"operationId":"reject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/outbound/{id}/approve":{"post":{"tags":["outbound-controller"],"operationId":"approve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/materials":{"get":{"tags":["material-controller"],"operationId":"list_2","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"lowStock","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultMaterial"}}}}}},"post":{"tags":["material-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Material"}}}}}}},"/api/materials/{id}/photo":{"post":{"tags":["material-controller"],"operationId":"uploadPhoto","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["photo"],"type":"object","properties":{"photo":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/inbound":{"get":{"tags":["inbound-controller"],"operationId":"list_3","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"materialId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultMapStringObject"}}}}}},"post":{"tags":["inbound-controller"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/users/profile/current":{"get":{"tags":["user-controller"],"operationId":"current","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserView"}}}}}}},"/api/reports/stock-by-category":{"get":{"tags":["report-controller"],"operationId":"stockByCategory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}}},"/api/reports/movement":{"get":{"tags":["report-controller"],"operationId":"movement","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":14}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}}},"/api/reports/low-stock":{"get":{"tags":["report-controller"],"operationId":"lowStock","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}}},"/api/reports/dashboard":{"get":{"tags":["report-controller"],"operationId":"dashboard","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/materials/categories/options":{"get":{"tags":["material-controller"],"operationId":"categories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"string"}}}}}}}}},"/api/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/docs":{"get":{"tags":["swagger-page-controller"],"operationId":"docs","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/docs/":{"get":{"tags":["swagger-page-controller"],"operationId":"docs_1","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/UserView"}}}}}}}},"/api/outbound/{id}":{"delete":{"tags":["outbound-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/inbound/{id}":{"delete":{"tags":["inbound-controller"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}}},"components":{"schemas":{"UserRequest":{"required":["name","username"],"type":"object","properties":{"username":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"status":{"type":"integer","format":"int32"}}},"UserView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"},"status":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"MaterialRequest":{"required":["category","code","name","unit"],"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"category":{"type":"string"},"spec":{"type":"string"},"unit":{"type":"string"},"supplier":{"type":"string"},"location":{"type":"string"},"safetyStock":{"type":"number","format":"double"},"remark":{"type":"string"}}},"Material":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"category":{"type":"string"},"spec":{"type":"string"},"unit":{"type":"string"},"supplier":{"type":"string"},"location":{"type":"string"},"safetyStock":{"type":"number","format":"double"},"stock":{"type":"number","format":"double"},"photoUrl":{"type":"string"},"remark":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"OutboundRequest":{"required":["materialId","quantity","recipient"],"type":"object","properties":{"materialId":{"type":"integer","format":"int64"},"quantity":{"type":"number","format":"double"},"recipient":{"type":"string"},"purpose":{"type":"string"},"remark":{"type":"string"}}},"ApprovalRequest":{"type":"object","properties":{"remark":{"type":"string"}}},"InboundRequest":{"required":["materialId","quantity"],"type":"object","properties":{"materialId":{"type":"integer","format":"int64"},"quantity":{"type":"number","format":"double"},"unitPrice":{"type":"number","format":"double"},"source":{"type":"string"},"remark":{"type":"string"}}},"RegisterRequest":{"required":["email","name","password"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/UserView"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"PageResultUserView":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/UserView"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"PageResultMapStringObject":{"type":"object","properties":{"items":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"PageResultMaterial":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Material"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}